+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Short time fourier transform
 - Spectogram extension trait
 - Preemphasis extension trait
+- Added Power and RMS for signal in the time domain
 - [Internal] Criterion for benchmarking and a benchmark for the STFT
 - [Internal] cargo-fuzz test for STFT that takes a corpus of wav files
 - [Internal] equivalence tests based on librosa implementations (ignored by default)